HRCPP-279 Refactor native test suite for C# client
