+++ build.bat
@@ -1,4 +1,7 @@
-set HOTRODCPP_HOME=%checkoutDir%/cpp-client/build_win/_CPack_Packages/WIN-x86_64/ZIP/infinispan-hotrod-cpp-%cppTag%-WIN-x86_64
+if [%HOTRODCPP_HOME%] == [] set HOTRODCPP_HOME=%checkoutDir%/cpp-client/build_win/_CPack_Packages/WIN-x86_64/ZIP/infinispan-hotrod-cpp-%cppTag%-WIN-x86_64
+echo Using HOTRODCPP_HOME=%HOTRODCPP_HOME%
+if [%generator%] == [""] set generator="Visual Studio 14 2015 Win64"
+echo Using generator -G %generator%
 
 rmdir /s /q build_windows
 mkdir build_windows
@@ -7,7 +10,7 @@ cd build_windows
 set "buildTest=%~1"
 
 call :unquote u_generator %generator%
-cmake -G "%u_generator%" -DHOTRODCPP_HOME=%HOTRODCPP_HOME%  -DPROTOBUF_INCLUDE_DIR=%PROTOBUF_INCLUDE_DIR% -DPROTOBUF_PROTOC_EXECUTABLE_CS=%PROTOBUF_PROTOC_EXECUTABLE_CS% -DGOOGLE_PROTOBUF_NUPKG=%GOOGLE_PROTOBUF_NUPKG% -DJBOSS_HOME=%JBOSS_HOME% -DIKVM_CUSTOM_BIN_PATH=%IKVM_CUSTOM_BIN_PATH% -DOPENSSL_ROOT_DIR=%OPENSSL_ROOT_DIR% %~4 ..
+cmake -G "%u_generator%" -DHOTRODCPP_HOME=%HOTRODCPP_HOME% -DPROTOBUF_PROTOC_EXECUTABLE_CS="%PROTOBUF_PROTOC_EXECUTABLE_CS%" -DGOOGLE_PROTOBUF_NUPKG="%GOOGLE_PROTOBUF_NUPKG%" -DPROTOBUF_INCLUDE_DIR=%PROTOBUF_INCLUDE_DIR% -DJBOSS_HOME=%JBOSS_HOME% -DIKVM_CUSTOM_BIN_PATH=%IKVM_CUSTOM_BIN_PATH% -DOPENSSL_ROOT_DIR=%OPENSSL_ROOT_DIR% -DCONFIGURATION=RelWithDebInfo %~4 ..
 if %errorlevel% neq 0 goto fail
 
 cmake --build . --config RelWithDebInfo
