--- nativeTestSuite/nativeTestSuite.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace nativeTestSuite
-{
-    class NativeTestSuite
-    {
-        public static void Main(string[] args)
-        {
-            AsyncTest.testMain(args);
-        }
-    }
-}
