+++ src/main/resources/proto/infinispan/hotrod/protobuf/base_types.proto
