+++ src/main/resources/proto/org/infinispan/protostream/message-wrapping.proto
