+++ src/main/resources/proto/org/infinispan/query/remote/client/query.proto
