+++ src/test/cs/Infinispan/HotRod/DefaultCacheForceReturnValueTest.cs
@@ -1,16 +1,21 @@
 using Infinispan.HotRod;
+using Infinispan.HotRod.Config;
+using Infinispan.HotRod.Tests.Util;
 using NUnit.Framework;
 using System;
 
 namespace Infinispan.HotRod.Tests
 {
-    public class DefaultCacheForceReturnValueTest : SingleServerAbstractTest
+    public class DefaultCacheForceReturnValueTest
     {
         private IRemoteCache<String, String> cache;
 
-        [SetUp]
-        public void BeforeTest()
+        [TestFixtureSetUp]
+        public void BeforeClass()
         {
+            ConfigurationBuilder conf = new ConfigurationBuilder();
+            conf.AddServer().Host("127.0.0.1").Port(11222).ConnectionTimeout(90000).SocketTimeout(6000);
+            RemoteCacheManager remoteManager = new RemoteCacheManager(conf.Build(), true);
             cache = remoteManager.GetCache<String, String>(true);
         }
