+++ src/test/cs/Infinispan/HotRod/Util/HotRodServer.cs
@@ -1,58 +1,61 @@
-﻿using Infinispan.HotRod.Impl;
-using Infinispan.HotRod.Config;
-using Infinispan.HotRod;
-using NUnit.Framework;
+﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Text;
-using System.Threading;
-using System;
+using System.Threading.Tasks;
+using NUnit.Framework;
+using Infinispan.HotRod.Impl;
+using Infinispan.HotRod.Config;
+using Infinispan.HotRod;
+using System.Diagnostics;
+using System.IO;
 
-namespace Infinispan.HotRod.Tests
+namespace Infinispan.HotRod.Tests.Util
 {
-    [TestFixture]
-    public abstract class SingleServerAbstractTest
+    class HotRodServer
     {
+        public const string hostname = "127.0.0.1";
+        public const int port = 11222;
+        string configurationFile;
         private Process hrServer;
 
-        protected Configuration conf;
-
-        protected RemoteCacheManager remoteManager;
-
-        [TestFixtureSetUp]
-        public void BeforeClass()
+        public HotRodServer(string configurationFile)
         {
-            try {
-                StartHotrodServer();
-            } catch (Exception ex) {
+            this.configurationFile = configurationFile;
+        }
+
+        public void StartHotRodServer()
+        {
+            try
+            {
+                Console.WriteLine("Starting Infinispan Server ...");
+                StartHotrodServerInternal();
+            }
+            catch (Exception ex)
+            {
                 ShutDownHotrodServer();
                 Console.WriteLine(ex.StackTrace);
                 throw ex;
             }
         }
 
-        public void StartHotrodServer()
+        void StartHotrodServerInternal()
         {
-            ConfigurationBuilder builder = new ConfigurationBuilder();
-            conf = builder.AddServers("127.0.0.1:11222").Build();
-
             string jbossHome = System.Environment.GetEnvironmentVariable("JBOSS_HOME");
             if (jbossHome == null)
             {
                 throw new Exception("JBOSS_HOME env variable not set.");
             }
 
-            Assert.IsTrue(PortProbe.IsPortClosed(conf.Servers()[0].Host(),
-                                                 conf.Servers()[0].Port(),
-                                                 millisTimeout:10000),
+            Assert.IsTrue(PortProbe.IsPortClosed(hostname, port, millisTimeout: 10000),
                           "Another process already listening on the same ip/port.");
 
             hrServer = new Process();
             hrServer.StartInfo.FileName = buildStartCommand(jbossHome);
+            hrServer.StartInfo.Arguments = "-c " + configurationFile;
             hrServer.StartInfo.UseShellExecute = false;
-            if (PlatformUtils.isUnix()) {
+            if (PlatformUtils.isUnix())
+            {
                 // Drop the output generated by the server on the console (data present in log file).
                 hrServer.StartInfo.RedirectStandardOutput = true;
                 hrServer.StartInfo.RedirectStandardError = true;
@@ -61,37 +64,30 @@ namespace Infinispan.HotRod.Tests
             }
             hrServer.Start();
 
-            Assert.IsTrue(PortProbe.IsPortOpen(conf.Servers()[0].Host(),
-                                               conf.Servers()[0].Port()),
+            Assert.IsTrue(PortProbe.IsPortOpen(hostname, port),
                           "Server not listening on the expected ip/port.");
-
-            remoteManager = new RemoteCacheManager(conf);
-            remoteManager.Start();
         }
 
-        [TestFixtureTearDown]
         public void ShutDownHotrodServer()
         {
-            if (remoteManager != null && remoteManager.IsStarted()) {
-                remoteManager.Stop();
-                Assert.IsFalse(remoteManager.IsStarted());
-            }
-
-            if (hrServer != null) {
+            Console.WriteLine("Shutting down Infinispan Server ...");
+            if (hrServer != null)
+            {
                 PlatformUtils.killServer(hrServer);
 
-                Assert.IsTrue(PortProbe.IsPortClosed(conf.Servers()[0].Host(),
-                                                     conf.Servers()[0].Port(),
-                                                     millisTimeout:10000),
+                Assert.IsTrue(PortProbe.IsPortClosed(hostname, port, millisTimeout: 10000),
                               "A process is still listening on the ip/port. Kill failed?");
             }
         }
 
         private string buildStartCommand(string homePath)
         {
-            if (PlatformUtils.isUnix()) {
+            if (PlatformUtils.isUnix())
+            {
                 return Path.Combine(homePath, "bin/standalone.sh");
-            } else {
+            }
+            else
+            {
                 return Path.Combine(homePath, "bin\\standalone.bat");
             }
         }
