+++ src/test/cs/Infinispan/HotRod/Util/PortProbe.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Net;
 using System.Net.Sockets;
 
-namespace Infinispan.HotRod.Tests
+namespace Infinispan.HotRod.Tests.Util
 {
     public class PortProbe
     {
