+++ src/test/resources/client-ca.pem
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDRDCCAiygAwIBAgIEVxdEnzANBgkqhkiG9w0BAQsFADBgMQswCQYDVQQGEwJ3
+dzEOMAwGA1UECAwFd29ybGQxDzANBgNVBAcMBnJlZGhhdDEOMAwGA1UECgwFamJv
+c3MxEzARBgNVBAsMCmluZmluaXNwYW4xCzAJBgNVBAMMAmNhMB4XDTE2MDQyMDA4
+NTgyNVoXDTE3MDQyMDA4NTgyNVowaDELMAkGA1UEBhMCd3cxDjAMBgNVBAgMBXdv
+cmxkMQ8wDQYDVQQHDAZyZWRoYXQxDjAMBgNVBAoMBWpib3NzMRMwEQYDVQQLDApp
+bmZpbmlzcGFuMRMwEQYDVQQDDApjbGllbnQtb25lMIIBIjANBgkqhkiG9w0BAQEF
+AAOCAQ8AMIIBCgKCAQEArqmHU6unZSYB3pWQaM4J29JKUqfBwFyDOqNeH4Ampeb+
+z4dvrnupUZFXXIOTndLUBxDMnZs2k94ufBblEPYl0yildXR8Z2iMGzAwhEc5a9De
+pGUwkXrOUR1rDNca830vGW4Rl4o7yVcDrgYdpOqeBk8MrX5tfuh1uXOf51hwlqnw
+KjUv6djdpLlIXuiRRZxz8KGx0n0Um33n0M7/Z8qionAii2x2dGbk3bhVsCbNPwSa
+DTGQi8XmAcYy5XsLNe/hWAoGhWNixVJZJbVXGgBJmbLCN/lvlICITwaj7Pm+rOym
+rBaPWAkVi4VG2xf+eF+DuvPqQlF912fY6bepvt0F3QIDAQABMA0GCSqGSIb3DQEB
+CwUAA4IBAQB0LnZY1+AiNNpoUcti3tNOwHnYGS6UY+wOWK9IUWQg7VPAGkiQw+My
+gS28ymTEWu7PTTX/QeSAqAitRO5fMuH1s9W5VXfVBhP+SG8eGGp5w5sKCsF/bPlj
+HQA6GDR7NPP7xkwrOIhoQLLwIAjWakCpWAFkw+M1tOY+SEv8vB2mzkaEmdHX7Zyk
+8dpDzP732CsTAe0Ee4VVMvzZgUiUvyOWdtBtWpb7g51KvOQ+/GfTBN8zjIuu+Hrv
+XVCddEFk9X4P/7+mNPPCauTAYOp819rUMwS8bQGj/rHNGcdbb99BIpuEDP3HmcJE
+RdOrficijsSnDkBY7t7NaJI9FF9vT7vk
+-----END CERTIFICATE-----
