+++ src/test/resources/proto3/bank.proto
