+++ templates/nunit.bat.in
@@ -1,10 +1,12 @@
 set TEST_ASSEMBLY_DIR=%1
+set SUITE=%2
+set OUTPUT_XML=%3
 
 set ORIG_PATH=%PATH%
 set PATH=${UT_PATH}
 set JBOSS_HOME=${HOTROD_JBOSS_HOME}
 
-"${NUNIT_NATIVE_BINARY}" %TEST_ASSEMBLY_DIR%hotrodcs-tests.dll
+"${NUNIT_NATIVE_BINARY}" %TEST_ASSEMBLY_DIR%hotrodcs-tests.dll /fixture=%SUITE% /result=%OUTPUT_XML%
 if %ERRORLEVEL% NEQ 0 goto FAIL
 goto SUCCESS
